%  Copyright (C) 2015 Christian Margreitter
\name{MDplot-internal}
\alias{parse_arguments}
\alias{testRequired}
\alias{getValue}
\alias{getFiles}
\alias{testAllowed}
\alias{getListOfKeys}
\alias{isKeySet}
\alias{integrate_curve}
\alias{split_equidistant}
\alias{plot_segments}
\alias{get_sign_digits}
\alias{fill_bins}
\alias{print_help}
\alias{setNumberDigits}
\alias{load_dssp_GROMACS}
\alias{load_XPM}
\alias{load_clusters_GROMACS}
\alias{find_Nth_occurrence}
\alias{split_GROMACS_atomnames}
\alias{split_AMBER_atomnames}
\alias{load_clusters_GROMACS}
\title{Internal functions for \code{MDplot}}
\description{Functions in \code{MDplot} not to be used by the user directly.
}
\author{Christian Margreitter}
\details{Not to be called by the user.}
\keyword{internal}
