% (C) 2015 Christian Margreitter
% last update: 2017-02-21
\name{MDplot_argument-class}
\docType{class}
\alias{MDplot_argument}
\alias{MDplot_argument-class}
\title{Arguments for bash script interface}
\description{
Container for bash arguments with "key" <> "value" pairs.
}
\section{Slots}{
  \describe{
    \item{\code{key}:}{Object of class \code{"character"}.}
    \item{\code{value}:}{Object of class \code{"character"}.}
  }
}
\author{Christian Margreitter}
\keyword{MDplot_argument}
