%  Copyright (C) 2015 Christian Margreitter

\name{MDplot_DSSP_summary}
\docType{function}
\alias{DSSP_summary}
\alias{DSSP}
\title{DSSP summary plot for secondary structure elements (proteins)}
\description{Plots summary plot for secondary structure motifs based on the output of the widely used classification program DSSP, which takes hydrogen bonds into account.}
\usage{MDplot_DSSP_summary( TABLE_datainput, BOOL_printLegend = FALSE, COLOURS_DSSP_summary = NULL, VEC_showValues = NULL, VEC_show_residues = NULL )}
\arguments
{
  \item{TABLE_datainput}{Table containing the information on the hydrogen bonds.}
  \item{.}{...}
  \item{.}{...}
  \item{.}{...}
  \item{.}{...}
}
\author{Christian Margreitter}
\keyword{DSSP}
