%  Copyright (C) 2015 Christian Margreitter
\name{MDplot_DSSP_timeseries}
\alias{MDplot_DSSP_timeseries}
\title{DSSP timeseries plot for secondary structure elements (proteins)}
\description{
Plots timeseries plot for secondary structure motifs in the context of the widely used DSSP algorithm. The default order is: "3-Helix", "4-Helix", "5-Helix", "Bend", "Beta-Bridge", "Beta-Strand", "Turn".}
\usage{
MDplot_DSSP_timeseries( LIST_timeseries,
                        BOOL_printLegend = TRUE,
                        VEC_timeBoundaries = NULL,
                        VEC_residueBoundaries = NULL,
                        BOOL_printNanoseconds = FALSE,
                        REAL_snapshotsPerNS = 1000,
                        ... )}
\arguments{
  \item{LIST_timeseries}{List construct consisting of lists, which are composed of a "name" (string) and a "values" table (x ... snapshots, y ... residues). Can be generated by \link{MDplot_load_DSSP_timeseries}.}
  \item{BOOL_printLegend}{If "TRUE", a legend is printed on the right hand side of the plot.}
  \item{VEC_timeBoundaries}{A vector of boundaries for the time in snapshots (two elements).}
  \item{VEC_residueBoundaries}{A vector of boundaries for the residues (two elements).}
  \item{BOOL_printNanoseconds}{If "TRUE", the snapshots are transformed into the respective nanoseconds, in dependence of parameter \code{REAL_snapshotsPerNS}.}
  \item{REAL_snapshotsPerNS}{Required, when \code{BOOL_printNanoseconds} is set to "TRUE" and gives the numbers of snapshots in the input file, that make up one nanosecond.}
  \item{...}{Graphical parameters.}
}
\examples{
# load table and return it into plotting function "MDplot_DSSP_timeseries()"
MDplot_DSSP_timeseries( MDplot_load_DSSP_timeseries( system.file( "extdata", package = "MDplot" ) ) )
}
\author{Christian Margreitter}
\keyword{DSSP}