%  Copyright (C) 2015 Christian Margreitter
\name{MDplot_XRMSD}
\docType{methods}
\alias{MDplot_XRMSD}
\title{XRMSD plot in heatmap style}
\description{Plots an heatmap according to the RMSD value for a 2D snapshot matrix, based on trajectories.}
\usage{
MDplot_XRMSD <- function( MAT_values,
                          BOOL_printLegend = TRUE,
                          VEC_xAxisRange = NULL,
                          VEC_yAxisRange = NULL,
                          VEC_colourRange = NULL,
                          ... )}
\arguments{
  \item{MAT_values}{Input matrix (three rows: x-values, y-values, RMSD-value). Can be generated by function \link{MDplot_load_XRMSD}.}
  \item{BOOL_printLegend}{If "TRUE", a legend is printed on the right hand side of the plot.}
  \item{VEC_xAxisRange}{A vector of boundaries for the x-snapshots (two elements).}
  \item{VEC_yAxisRange}{A vector of boundaries for the y-snapshots (two elements).}
  \item{VEC_colourRange}{A vector with colour names for the span palette.}
  \item{...}{Graphical parameters.}
}
\examples{
\dontrun{
# load matrix and return it into plotting function "MDplot_XRMSD()"
# see "extdata/MDplot_XRMSD_example.txt" for format information
#MDplot_XRMSD( MDplot_load_XRMSD( system.file( "extdata/MDplot_XRMSD_example.txt", package = "MDplot" ) ) )
}}
\author{Christian Margreitter}
\keyword{XRMSD}