%  Copyright (C) 2015 Christian Margreitter
\name{MDplot_load_DSSP_timeseries}
\alias{MDplot_load_DSSP_timeseries}
\title{DSSP timeseries plot loading}
\description{
Loads DSSP output files from a specified directory and combined it into a list construct suitable for \link{MDplot_DSSP_timeseries}.}
\usage{
MDplot_load_DSSP_timeseries( STRING_folder )}
\arguments{
  \item{STRING_folder}{Folder, in which the DSSP output files are located.}
}
\author{Christian Margreitter}
\keyword{DSSP}