%  Copyright (C) 2015 Christian Margreitter
\name{MDplot_XRMSD}
\docType{methods}
\alias{MDplot_XRMSD}
\title{XRMSD plot in heatmap style}
\description{Plots an heatmap according to the RMSD value for a 2D snapshot matrix, based on trajectories.}
\usage{
xrmsd( xrmsdValues,
       printLegend = TRUE,
       xaxisRange = NA,
       yaxisRange = NA,
       colours = NA,
       ... )}
\arguments{
  \item{xrmsdValues}{Input matrix (three rows: x-values, y-values, RMSD-value). Can be generated by function \link{MDplot::load_xrmsd}.}
  \item{printLegend}{If "TRUE", a legend is printed on the right hand side of the plot.}
  \item{xaxisRange}{A vector of boundaries for the x-snapshots (two elements).}
  \item{yaxisRange}{A vector of boundaries for the y-snapshots (two elements).}
  \item{colours}{A vector with colour names for the span palette.}
  \item{...}{Graphical parameters.}
}
\examples{
\dontrun{
# see "extdata/xrmsd_example.txt" for format information
MDplot::rmsd( MDplot::load_xrmsd( system.file( "extdata/xrmsd_example.txt", package = "MDplot" ) ) )
}}
\author{Christian Margreitter}
\keyword{XRMSD}